fix: 계산 자료형 double 변경

+++ src/main/java/calculator/Application.java
@@ -8,7 +8,7 @@ public class Application {
 
     public static void main(String[] args) {
         List<String> formula = FormulaInput.getFormula();
-        int result = Calculator.run(formula);
+        double result = Calculator.run(formula);
         System.out.println(result);
     }
 }
