+++ src/main/java/calculator/service/Calculator.java
@@ -9,18 +9,18 @@ public class Calculator {
     private Calculator() {
     }
 
-    public static int run(List<String> formula){
+    public static double run(List<String> formula){
         Stack<String> formulaStack = new Stack<>();
 
         for(int i = formula.size()-1; i>=0; i--){
             formulaStack.push(formula.get(i));
         }
         while(formulaStack.size() > 1){
-            int prev = Integer.parseInt(formulaStack.pop());
+            double prev = Double.parseDouble(formulaStack.pop());
             Operation operator = Operation.fromString(formulaStack.pop()).get();
-            int next = Integer.parseInt(formulaStack.pop());
+            double next = Double.parseDouble(formulaStack.pop());
 
-            int result = 0;
+            double result = 0;
             if(Operation.PLUS == operator){
                 result = add(prev, next);
             }else if(Operation.MINUS == operator){
@@ -33,23 +33,22 @@ public class Calculator {
             formulaStack.push(String.valueOf(result));
         }
 
-        return Integer.parseInt(formulaStack.pop());
+        return Double.parseDouble(formulaStack.pop());
     }
 
-    public static int add(int prev, int next) {
+    public static double add(double prev, double next) {
         return prev + next;
     }
 
-    public static int sub(int prev, int next) {
+    public static double sub(double prev, double next) {
         return prev - next;
     }
 
-    public static int multiply(int prev, int next) {
+    public static double multiply(double prev, double next) {
         return prev * next;
     }
 
-    // TODO: 실수형 고려
-    public static int divide(int prev, int next) {
+    public static double divide(double prev, double next) {
         return prev / next;
     }
 }
